move static methods to helper class
